Move AcademyContainer in containers/academy
