+++ src/containers/academy/index.ts
@@ -2,8 +2,8 @@ import { connect, MapDispatchToProps } from 'react-redux'
 import { withRouter } from 'react-router'
 import { bindActionCreators, Dispatch } from 'redux'
 
-import { changeToken, fetchUsername } from '../actions/session'
-import Academy, { IDispatchProps } from '../components/academy'
+import { changeToken, fetchUsername } from '../../actions/session'
+import Academy, { IDispatchProps } from '../../components/academy'
 
 const mapDispatchToProps: MapDispatchToProps<IDispatchProps, {}> = (dispatch: Dispatch<any>) =>
   bindActionCreators(
